pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["artec*"]
exclude = ["img*"]

[project]
name = "Artec"
version = "0.3.0"
authors = [
  { name="HushmKun", email="HushmKun@outlook.com" },
  { name="Link-", email="bsm.dgdy@gmail.com" },
]
description = "Creates a configurable python project template in a given directory."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
        "Intended Audience :: Developers",
        "Development Status :: 4 - Beta",
        "Environment :: Console",
        "Programming Language :: Python :: 3",
        "Topic :: Software Development :: Build Tools"
      ]
dependencies = [
    "GitPython"
]

[project.license]
file = "LICENSE"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
]
test = [
    "pytest",
]


[project.urls]
"Homepage" = "https://github.com/HushmKun/Artec"
"Bug Tracker" = "https://github.com/HushmKun/Artec/issues"

[project.scripts]
artec = "artec.__main__:main"
